% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_func.R
\name{envelope}
\alias{envelope}
\title{Envelope function}
\usage{
envelope(x = 4, y = 4, ...)
}
\arguments{
\item{number}{of children in y family}
}
\description{
This is a function that helps you decide how much envelope money each child should receive in each family.
}
\details{
the function will print how much envelope money should each child receive in each family.
}
