\name{fiberpaper}
\alias{fiberpaper}
\docType{data}
\title{Pulp and Paper Data
}
\description{Pulp and paper property}
\usage{data("fiberpaper")}
\format{
  A data frame with 62 observations on the following 8 variables.
  \describe{
    \item{\code{V1}}{Breaking length.}
    \item{\code{V2}}{Elastic modulus.}
    \item{\code{V3}}{Stress at failure.}
    \item{\code{V4}}{Burst strength.}
    \item{\code{V5}}{Arithmetic fiber length.}
    \item{\code{V6}}{Long fiber fraction.}
    \item{\code{V7}}{Fine fiber fraction.}
    \item{\code{V8}}{Zero span tensile.}
  }
}
\details{
     This data set contains measurements of properties of pulp fibers
     and the paper made from them.
     }
\references{
     Johnson, R.A. and Wichern, D.W. (2007). Applied Multivariate
     Statistical Analysis, 6th edition.
     }

\keyword{datasets}
