% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_func.R
\name{envelope}
\alias{envelope}
\title{Envelope function
This is a function that helps you decide how much envelope money each child
should receive in each family reasonably and fairly.}
\usage{
envelope(x = 4, y = 4, ...)
}
\arguments{
\item{x}{number of children in x family}

\item{y}{number of children in y family}

\item{...}{number of more children in other families}
}
\value{
a vector of envelope money for x and y family
}
\description{
Envelope function
This is a function that helps you decide how much envelope money each child
should receive in each family reasonably and fairly.
}
\details{
The function will print the reasonable amoung of envelope money
should each child receive in each family.
}
